Refactor Navbar -> NavbarSimple and CustomNavbar -> Navbar
