+++ docs/components_page/components/navbar/__init__.py
@@ -10,14 +10,14 @@ from ...helpers import (
     load_source_with_environment,
 )
 from ...metadata import get_component_metadata
-from .navbar import navbar
+from .simple import navbar as navbar_simple
 
-GITHUB_EXAMPLES = "https://github.com/ASIDataScience/dash-bootstrap-components/blob/master/examples/advanced-component-usage/CustomNavbar.py"  # noqa
+GITHUB_EXAMPLES = "https://github.com/ASIDataScience/dash-bootstrap-components/blob/master/examples/advanced-component-usage/Navbars.py"  # noqa
 
 HERE = Path(__file__).parent
 
+navbar_simple_source = (HERE / "simple.py").read_text()
 navbar_source = (HERE / "navbar.py").read_text()
-custom_navbar_source = (HERE / "custom_navbar.py").read_text()
 
 
 def get_content(app):
@@ -26,38 +26,38 @@ def get_content(app):
         html.P(
             dcc.Markdown(
                 "There are two navbar components in "
-                "*dash-bootstrap-components*, `Navbar` and `CustomNavbar`. "
-                "`Navbar` is simpler but less flexible, whereas "
-                "`CustomNavbar` is fully customisable, but requires more "
+                "*dash-bootstrap-components*, `NavbarSimple` and `Navbar`. "
+                "`NavbarSimple` is simpler but less flexible, whereas "
+                "`Navbar` is fully customisable, but requires more "
                 "boilerplate to get working."
             )
         ),
-        html.H4("Navbar"),
+        html.H4("NavbarSimple"),
         html.P(
             dcc.Markdown(
-                "The pre-built `Navbar` consists of a 'brand' on the left, to "
-                "which you can attach a link with `brand_href`, and a number "
-                "nav items as its children. You can use a mix of `NavItem` "
-                "and `DropdownMenu`."
+                "The pre-built `NavbarSimple` consists of a 'brand' on the "
+                "left, to which you can attach a link with `brand_href`, and "
+                "a number nav items as its children. You can use a mix of "
+                "`NavItem` and `DropdownMenu`."
             )
         ),
         html.P(
             dcc.Markdown(
-                "The `Navbar` will collapse on smaller screens, and add a "
-                "toggle for revealing the navigation items. You can modify "
+                "The `NavbarSimple` will collapse on smaller screens, and add "
+                "a toggle for revealing the navigation items. You can modify "
                 "which breakpoint the expansion happens at with the `expand` "
                 "keyword argument. The default is `'md'`."
             )
         ),
-        ExampleContainer(navbar),
-        HighlightedSource(navbar_source),
-        html.H4("Custom Navbar"),
+        ExampleContainer(navbar_simple),
+        HighlightedSource(navbar_simple_source),
+        html.H4("Navbar"),
         html.P(
             dcc.Markdown(
                 "If you want to have more control over the layout of your "
-                "navbar you can use the `CustomNavbar` component. This gives "
-                "you full control over the children, but you will have to "
-                "write your own callbacks to achieve things like the toggle "
+                "navbar you can use the `Navbar` component. This gives you "
+                "full control over the children, but you will have to write "
+                "your own callbacks to achieve things like the toggle "
                 "behaviour on small screens."
             )
         ),
@@ -71,18 +71,16 @@ def get_content(app):
             ]
         ),
         ExampleContainer(
-            load_source_with_environment(
-                custom_navbar_source, "navbar", {"app": app}
-            )
+            load_source_with_environment(navbar_source, "navbar", {"app": app})
         ),
-        HighlightedSource(custom_navbar_source),
+        HighlightedSource(navbar_source),
+        ApiDoc(
+            get_component_metadata("src/components/nav/NavbarSimple.js"),
+            component_name="NavbarSimple",
+        ),
         ApiDoc(
             get_component_metadata("src/components/nav/Navbar.js"),
             component_name="Navbar",
-        ),
-        ApiDoc(
-            get_component_metadata("src/components/nav/CustomNavbar.js"),
-            component_name="CustomNavbar",
         ),
         ApiDoc(
             get_component_metadata("src/components/nav/NavbarBrand.js"),
