+++ docs/demos/demo_layout.py
@@ -3,7 +3,7 @@ import dash_core_components as dcc
 import dash_html_components as html
 from dash import Dash
 
-_navbar = dbc.Navbar(
+_navbar = dbc.NavbarSimple(
     brand="Demo",
     brand_href="#",
     sticky="top",
