+++ docs/templates/index.html
@@ -62,7 +62,7 @@ import dash_bootstrap_components as dbc
 import dash_core_components as dcc
 import dash_html_components as html
 
-navbar = dbc.Navbar(
+navbar = dbc.NavbarSimple(
     children=[
         dbc.NavItem(dbc.NavLink("Link", href="#")),
         dbc.DropdownMenu(
