+++ setup.cfg
@@ -9,7 +9,7 @@ exclude =
     docs/components_page/components/input_group/button.py,
     docs/components_page/components/input_group/dropdown.py,
     docs/components_page/components/list_group/links.py,
-    docs/components_page/components/navbar/custom_navbar.py,
+    docs/components_page/components/navbar/navbar.py,
     docs/components_page/components/popover/popover.py,
     docs/components_page/components/progress/progress.py,
     docs/components_page/components/table/kwargs.py,
